dj_project_1/app1/viewsParcel.py
from rest_framework import viewsets
from rest_framework.permissions import IsAuthenticated, AllowAny
from rest_framework.authentication import TokenAuthentication
from .serializers import ParcelSerializer
from.models import Parcel

class ViewsParcel(viewsets.ModelViewSet):
    serializer_class = ParcelSerializer
    queryset = Parcel.objects.all()
    authentication_classes = [TokenAuthentication]

    def get_permissions(self):
        if self.action in ['list', 'retrieve']:
            self.permission_classes = [AllowAny]
        else:
            self.permission_classes = [IsAuthenticated]
        return super().get_permissions()